Implement dictionary bar for scan popup.
This dictionary bar operates completely separately from the dictionary bar in
the main window, and has its own set of muted dictionaries. The bar is on the
right of the window by default, but can be moved once the window is pinned
down.

Windows users, please test, this may have some Windows-specific problems.