Implemented Dictionaries Pane, to show dicts in which results were found.
It is a separate detacheable pane, which could also be enabled/disabled,
it tracks the current tab's contains, and lists those dictionaries that
contributed their content to the current translation.

Clicking on dictionary name jumps to the appropriate article from
this dictionary, which improves the navigation flow.

The Dictionaries Pane:
  * Shows both dictionaries' names and icons.
  * Does not consume resources when hidden.
  * Properly styled.
  * Properly handles app-wide shortcuts
    (Esc, Ctrl-L, Alt-D, Enter, Alt-Up/Down).