Restored the code to add actions from the hidden menu back to the main window.
So that if Menu Button is hidden (or the whole toolbar is hidden) all
the actions would still work, including the one to show the main menu back.

One of the possibilities when the Menu Button is hidden is when not all
toolbar buttons fit into winow and are being automatically hidden.