Use the same window flags for the scan window, on Win and Linux.
Previously, the scan popup used Qt::Popup flag on Linux, but this
flag is incompatible with the new translate box UI, esp. the
floating suggestion list, since in the Qt::Popup-styled windows
all the events go to the popup, hence the creation of additional
windows is problematic, and those new windows don't work properly
since they don't receive events.

Without this change, the scrolling of the suggestion list in the
Scan Popup's translate box is impossible.