On Windows x64, fixed #208: GoldenDict crashes when indexing huge dictionaries.
According to Microsoft documentation, on x64 systems the virtual address space
for 32-bit process can be extended from 2Gb to 4Gb if if the application is
compiled with the /LARGEADDRESSAWARE switch.

I verified that with this switch we get full 4Gb on x64 systems and the
indexing of huge dictionaries works just fine now.

On x32 systems we can't do much though.