Properly handle non-normalized unicode headwords
With that change users should be able to search headwords in
any form. For example:

U+03B5 GREEK SMALL LETTER EPSILON and U+0301 COMBINING ACUTE ACCENT

is considered equal to

U+03AD GREEK SMALL LETTER EPSILON WITH TONOS

And no matter in what form the headword is provided in the dictionary, users will be able to find it,
even using the different form.