Add --toggle-scan-popup command line option
This option allows to choose whether scan popup is enabled or disabled
on application start without changing the
"Start with scan popup turned on" option in Preferences.

More importantly, this option allows to assign a global keyboard
shortcut that toggles scan popup for the already running goldendict
instance in system preferences. So it implements #81 in a way.

The advantages of the command line option over a hard-coded shortcut:
  * each user can use the individually preferred key binding;
  * there is no hotkey conflict risk.

The disadvantages of the command line option are:
  * poor feature discoverability - the best I can think of is listing
    this option in the Command line switches section of Goldendict Help;
  * each user that needs the feature has to make an extra effort
    to assign a shortcut in system preferences.