Improve large buffer performance in AllowFrameReply with Qt5
An example of a large buffer can be found in the "Michael Jordan"
article from English Wikipedia. Search for "Listen to this article"
inside this article's text or turn on the Auto-pronounce option
in Preferences. After requesting the audio playback, start scrolling
the article to notice goldendict's UI freeze.

The difference in this case is quite noticeable on my GNU/Linux machine
in Release mode: after the audio file is downloaded and before
the playback starts, the UI freezes for 18 seconds
with QByteArray buffer, while the freeze is almost absent
with std::deque buffer. The UI freezes for about a second
with std::deque buffer in Debug mode, though this shouldn't be an issue.