Safety menu sequence in dictionaries order window
On accidentally mouse moving when right button clicking first context menu stroke may be apply. If in this stroke is dictionaries reordering it may be not noticed (especially, when filter is enabled), and after the windows OK button clicking ALL dictions will be reordered!
This commit places on context menu first stroke safety "Dictionary headwords".