Use absoluteFilePath instead of canonicalFilePath
handle symlink dict file properly.
previous way (canonicalFilePath) will dereference the symlink, thus cause
the following problem:
e.g. the directory structure is
  a/a.mdx -> ../1.mdx    # the main file of a '.mdx' dict
  a/a.mdd -> ../2.mdd
  a/a.css -> ../3.css    #  the css filename 'a.css' is hard-coded in '1.mdx'
  1.mdx
  2.mdd
  3.css
If you follow the symlink when openning dict 'a/a.mdx', goldendict will
read '1.mdx'. But, it will fail to find the corresponding '1.mdd' and
the hard-coded 'a.css'.