Update es_Es.ts
Hi! I'm a Spanish biologist, linguist and writer. GoldenDict is a wonderful program that is helping me a lot in my everyday work. This is my first contribution. I have done a full translation to Spanish of many untranslated words and a complete update of several grammatical errors that there was in the previous version of this file. The other Spanish variants also present grammatical errors that I can correct.

I'd be very glad if you could update the file with my changes.

Thanks in advance.