Add "Jump to Target Article" action
This new action is particularly useful when the "Automatically scroll to
target article" option is disabled. Even if the user prefers
translations from higher-priority dictionaries most of the time,
sometimes she may want to read the target article. This action allows to
quickly scroll to the target article via its Alt+T shortcut.