Reduce communication between JavaScript and C++
This is the fifth in a series of commits that get rid of
evaluateJavaScript() calls. The rationale is given in the first such
commit "Don't call gdCheckArticlesNumber() from C++ code".

This commit reduces the maximum number of interactions between
JavaScript and C++ code in ArticleView::setCurrentArticle() from 3 to 1
by getting rid of one evaluateJavaScript() call and a
gdArticleView.onJsActiveArticleChanged() call.

I have tried to emit an ArticleViewJsProxy's signal in place of the
remaining evaluateJavaScript() call. This made the C++ code cleaner. But
connecting to C++ signals in JavaScript code has two downsides:
1. The Qt WebKit version of GoldenDict crashes in the function connected
   to the signal. This can be worked around by performing otherwise
   unnecessary checks in the JavaScript function.
2. In the upcoming Qt WebEngine version of GoldenDict, gdArticleView is
   not available when a page starts to load. When
   ArticleView::setCurrentArticle() is called before gdArticleView
   becomes available, the emitted signal has no effect. This could be
   reproduced by selecting a dictionary in Results Navigation Pane right
   after a page begins loading.