Add CMakeLists.txt to add CMake build system support
This CMakeLists is not very perfect for that it just has a basic support for Linux, but it seems to work well with Qt 5.

Note that although this CMakeLists has support for build with Qt 4, it's a little difficult to verify if it works since many Linux distros drops Qt 4, and some codes in this project seem not to work with Qt 4 anymore.