Move tray icons to the XDG hicolor icon theme
Install the two icons used for the tray icon as part of the XDG hicolor
icon theme, so they are found by default without falling back to their
version as available from the embedded resources.

Since the parent commit, the following bug occurs on
freedesktop.org-compatible systems whenever the current icon theme does
not override the icons "goldendict-tray" and "goldendict-scan-tray": the
notification area (systray) icon is the same as the program icon, even
if the scan popup is enabled, which breaks the visual distinction
between the scan popup states. This commit fixes the regression by
providing the two searched-for icons and consequently disabling the use
of the less specific icon "goldendict".

The same bug affects icon themes that override the "goldendict" icon but
not the "goldendict-tray" and "goldendict-scan-tray" icons since
eea6f686f6abc4735acb3edfdd1a88112c7b5f2f (2018) started searching for
the two tray icons in the current icon theme. This old bug still remains
because the priority of a less specific icon from the current theme
("goldendict") is higher than that of an exact-match icon (e.g.
"goldendict-scan-tray") from a parent theme. Fixing this properly and
allowing icon themes to override only the application icon requires
renaming the two tray icons so that the icon search algorithm no longer
considers the application icon a (less specific) match. The new names
could be "goldendict.tray" and "goldendict.scantray", seeing as the dash
character is used to separate levels of specificity. See
https://specifications.freedesktop.org/icon-naming-spec/latest/#guidelines

Tweak the resources file to use the old names for the new paths, so
there is no need to change their references everywhere.