WebEngine-specific: allow opening URL with a fragment in a new tab
The local resource error occurs because when the window.location
assignment code is executed, the initial "gdlookup://localhost?blank=1"
page has not been loaded yet in the new tab. JavaScript navigation from
nowhere to a local-scheme URL is forbidden. Delaying the JavaScript call
by 100 ms also works, but why waste time loading the blank page?

Replace the "scrollto" query item value in order to set correct current
article in the Qt WebEngine version. In the Qt WebKit version, remove
the "scrollto" query item in order to scroll to the fragment ID rather
than to the top of the current article in the new foreground tab (this
is a related minor fix, so the commit is almost WebEngine-specific).