Don't wait for QWebView::loadFinished() to update the UI
Clear the dictionary list and disable the audio button when a page
starts loading. As soon as an article is loaded, append it to the
dictionary list and enable the audio button if the article contains the
first audio link on the page.

The UI and ArticleView's data members could be out of date for a long
time in case of slow-loading online dictionaries. Before the recent
commits that got rid of evaluateJavaScript() calls, ArticleView at least
used up-to-date JavaScript variables in response to explicit user
actions, such as activating an article or triggering the Pronounce Word
action. Though the state of the audio button and the dictionary list in
the Results Navigation Pane could remain out of date until the page
fully loaded.

The prompt UI updates introduced in this commit improve correctness by
ensuring the freshness of the data and enhance perceived responsiveness.

The audio link script embedded in articles no longer references the
dictionary ID => remove the unused dictionaryId parameter.

Move scripts above stylesheets on HTML pages. Running the scripts
earlier invokes ArticleView::onJsPageInitStarted() earlier, which speeds
up UI updates when a new page is loaded.