WebEngine-specific: synchronize current article updates
The current article variable is unique in that both C++ code and
JavaScript code can modify it and send the updated value to the other
side. This is not a problem in the Qt WebKit version where C++ and
JavaScript communicate via synchronous function calls. But in the Qt
WebEngine version, with its asynchronous messages that are passed
between processes, the current article value can end up wrong or even
inconsistent between C++ and JavaScript.

Introduce two Date JavaScript variables and two corresponding QDateTime
ArticleView data members. These two timestamps are passed back and forth
in messages that update the current article. Comparing stored and
received timestamp values allows to reject stale messages and keep the
fresher current article value.

For example, if the user clicks on a dictionary item in Results
Navigation Pane after a new page starts loading but before
ArticleView::onJsPageInitStarted() for the new page is invoked, the
JavaScript side receives a current article value that may even be
invalid on the new page. Even if the received current article value is
valid, it should be rejected, because user interaction with outdated UI
produced the value. If this value is not rejected and a valid
activeArticleIndex has been sent in the onJsPageInitStarted() message,
the JavaScript's and the C++'s current article values become
inconsistent. The page timestamp eliminates this issue by letting the
JavaScript side reject current article updates pertaining to another
(usually the previous) page.

Another example: the user clicks on a dictionary item in Results
Navigation Pane while a page is loading. At the same time, the current
article (gdCurrentArticle) is loaded and isActive == true is sent in
onJsArticleLoaded() message. The C++ side must reject JavaScript's
current article update to keep the user-selected value and to prevent
inconsistency with the JavaScript side, which is going to update its
current article once it receives an already sent
gdOnCppActiveArticleChanged() message. Comparing the two timestamp data
members of ArticleView allows to fix this issue.

I have managed to reproduce the two example issues above at the parent
commit. Introducing artificial message delivery delays (using timers)
allowed me to reproduce other timing issues and test all the code added
in this commit.

I have tried to emit an ArticleViewJsProxy's signal in place of the
gdOnCppActiveArticleChanged runJavaScript() call in the Qt WebEngine
version in order to avoid explicitly converting QDateTime to QString and
JavaScript String to Date. But this did not work as expected for some
reason: the timestamp values received by JavaScript were of the type
String. So I decided to call runJavaScript() rather than emit the signal
in order to prevent a quiet JavaScript type conversion regression in
case the issue is fixed and the type received by the JavaScript function
connected to the C++ signal becomes Date in a future Qt version.