Cancel data request in ArticleResourceReply::close()
From the documentation for QNetworkReply::close():
Closes this device for reading. Unread data is discarded

When ArticleResourceReply::close() is called, nothing else is going to
be read, so canceling the data request is safe. This should improve
performance by aborting no longer useful reading from disk and network
requests. close() is called, for example, when the user requests
translation of a different phrase or goes back/forward in web history
before the current page is fully loaded.

In the Qt WebKit version close() is never called. The reply is destroyed
immediately instead, which works just fine. Don't #ifdef the overriding
of close(), because it is correct in the Qt WebKit version too, if not
useful in practice.