Save Article: use QSet's hash instead of QCryptographicHash
The MD5 algorithm has a different purpose and is wholly unsuitable to
ensuring uniqueness of saved resources. The MD5 hash function is slower
than the QSet's, which is optimized for speed. And the MD5-based
implementation doesn't even guarantee correctness, because collisions
are not handled (probably dismissed as unlikely).

Hash the destination subpath instead of the URL: if the same subpath is
used for two downloaded resources, only one of them would end up saved.

Use QSet instead of std::set, because a hash table should be faster than
a red-black tree for keeping track of duplicates.

The performance impact of this change is not measurable in Qt 5.