Support activating an article while it is being loaded
The code didn't take this possibility into account before.

It is very easy to activate an article that is being loaded in the Qt
WebEngine version: just request a huge MediaWiki article, e.g. "Paris"
from English Wikipedia, and click on an empty place within once it
appears. Doing the same in the Qt 4 WebKit version is much more
difficult: click on an empty place within a huge article just before the
article becomes visible. I wasn't able to activate a not-fully-loaded
article in the Qt 5 WebKit version at all.

The wrong behavior after such an early activation before this commit:
* in the Qt WebEngine version no dictionary row is selected in Results
  Navigation Pane; the target article or fragment is scrolled to;
* in the Qt WebKit version no dictionary row is selected in Results
  Navigation Pane if the activated article is topmost; the topmost
  article's dictionary row remains selected in Results Navigation Pane
  otherwise.