Handle goldendict:// and dict:// URI schemes
Specifications:
https://specifications.freedesktop.org/desktop-entry-spec/desktop-entry-spec-latest.html
https://specifications.freedesktop.org/shared-mime-info-spec/shared-mime-info-spec-latest.html#idm45659000071008

GoldenDict now opens the supported URIs on GNU/Linux. However,
platform-specific changes may be necessary to make this work on other
operating systems. Hopefully the required changes can be implemented
without recompiling GoldenDict, similarly to the simple modification of
GoldenDict's desktop file in this commit. In this case, a user can make
the necessary changes and test locally, then create a pull request to
complete URI scheme support on his/her preferred OS.

This feature has been requested in #735, #1139 and #1280.
The implementation is based on #1624 and its review in #1625.