Allow building against libebu in place of libeb
This commit is based on the first commit of an unmerged porting pull
request #1344. As I wrote in a comment to that pull request, the first
step in porting from libeb to libebu is to allow optionally building
GoldenDict against libebu and let GNU/Linux distributions try switching
to libebu, if they feel like it. Then the distros' experience and bug
reports (or lack thereof) can help us decide whether to make libebu the
default dependency and whether to replace GoldenDict's bundled libeb
files with the corresponding libebu files in maclibs/ and winlibs/.