Increase minimum ArticleView zoom factor from 0.1 to 0.3
The former minimum zoom factor value is less than the minimum zoom
factor of QWebEngineView 0.25. The first two zooming-in steps starting
from that minimum value have no visible effect in the Qt WebEngine
version. This may scare a user into thinking that reducing zoom factor
to its minimum value cannot be undone. The new minimum is not 0.25,
because 1.0-0.25 is not divisible by 0.1 (the added comment explains why
it should be divisible).

Article text is still unreadable when zoom factor is set to its new
minimum value.