MDict: fix reading record blocks
Comparison operators for MdictParser::RecordIndex and qint64 use
RecordIndex's data members shadowStartPos and shadowEndPos. The binary
search in MdictParser::RecordIndex::bsearch() uses these comparison
operators to find a qint64 in an ordered RecordIndex collection.
The condition for a bsearch() in MdictParser::readRecordBlock() uses
RecordIndex's data member endPos in place of shadowEndPos by mistake.

Assignments a few lines below in readRecordBlock() confirm that this is
a mistake:
    RecordIndex const & recordIndex = recordBlockInfos_[idx];
...
      recordSize = recordIndex.shadowEndPos - i->first;
...
    recordInfo.recordSize = recordSize;
MdictParser::RecordInfo::recordSize must be non-negative because it is
used as a size. For example, in
IndexedMdd::loadFile(gd::wstring const &, std::vector<char> & result):
    result.resize( indexEntry.recordSize );
std::vector::resize() takes an unsigned size_type argument. Passing a
negative integer to it would attempt to resize to a huge size close to
the maximum value of std::vector::size_type.

There are two practical consequences of this mistake:
1. An affected image is never displayed in the article view.
2. GoldenDict occasionally crashes while loading an affected article.

I can consistently reproduce the crash by running a Debug build of
GoldenDict via GDB and loading the single article of the affected
dictionary attached to #1672. The backtrace:
    Thread 7 "Thread (pooled)" received signal SIGSEGV, Segmentation fault.
    #0  in  () at /usr/lib/libc.so.6
    #1  in Mdx::IndexedMdd::loadFile(std::__cxx11::basic_string<wchar_t, std::char_traits<wchar_t>, std::allocator<wchar_t> > const&, std::vector<char, std::allocator<char> >&)
        (this=0x55555639d5d0, name=L"\\OPEU4_0002.png", result=std::vector of length 1993913, capacity 1993913 = {...}) at ../mdx.cc:184
    #2  in Mdx::MddResourceRequest::run() (this=0x555556c15f90)
        at ../mdx.cc:849
    #3  in Mdx::MddResourceRequestRunnable::run()
        (this=0x555556c163d0) at ../mdx.cc:791
    #4  in  () at /usr/lib/libQt5Core.so.5
    #5  in  () at /usr/lib/libQt5Core.so.5
    #6  in  () at /usr/lib/libc.so.6
    #7  in  () at /usr/lib/libc.so.6

Correct the mistake in readRecordBlock() and increment the Mdict format
version, because the crash goes away only after reindexing an affected
dictionary.

Fixes #1672