Fix translations of "Portuguese" to Portuguese
The translation of "Portuguese" to Brazilian Portuguese has been
"Português brasileiro" for years. This was fine until a recent commit
added a European Portuguese translation and copied this particular
translation verbatim. So now there are two consecutive items with the
text "Português brasileiro" in the "Interface language" combobox in
GoldenDict Preferences if one of the Portuguese variants is the current
interface language.

Each language name in the "Interface language" combobox is translated to
the same current interface language. Therefore, the texts of combobox
items that represent different variants of the same language cannot be
different. So change both translations to "Português" and let the
national flags disambiguate the language variants as is the case for
other languages.